[Test] 깨지는 테스트 수정
